pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=5.0,<6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
    "whitenoise>=6.12.0",
]
